Started work. Stopped to study.
Started initial work on project.
Found old website to alter in order to use as website for app.
Stopped work because I decided to study for exams. Learning all of the
new material for apps and websites was interfering with my concentration
while studying so I decided to leave it until after exams.

Removed old website as it was the wrong one.